agentic_work/28_custom_runner/pyproject.toml
[project]
name = "28-custom-runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langfuse>=3.8.1",
    "openai-agents>=0.4.1",
    "openinference-instrumentation-openai-agents>=1.3.0",
]
